server/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
email-validator
asyncpg
httpx